% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SISforest.R
\name{SISforest}
\alias{SISforest}
\alias{SISforest.formula}
\alias{SISforest.default}
\title{Sure Independence Screening for Forest Structures}
\usage{
SISforest(X, ...)

\method{SISforest}{formula}(
  formula,
  data = NULL,
  criteria = c("DCor", "Pearson", "CvM", "MI", "HHG", "HSIC", "Gini")[1],
  ntrees = 100,
  storeOOB = TRUE,
  replacement = TRUE,
  stratify = TRUE,
  ratOOB = 0.368,
  parallel = TRUE,
  numCores = Inf,
  MaxDepth = Inf,
  numNode = Inf,
  MinLeaf = 10,
  subset = NULL,
  weights = NULL,
  na.action = na.fail,
  Xscale = "Min-max",
  ...
)

\method{SISforest}{default}(
  X,
  y,
  criteria = c("DCor", "Pearson", "CvM", "MI", "HHG", "HSIC", "Gini")[1],
  ntrees = 100,
  storeOOB = TRUE,
  replacement = TRUE,
  stratify = TRUE,
  ratOOB = 0.368,
  parallel = TRUE,
  numCores = Inf,
  MaxDepth = Inf,
  numNode = Inf,
  MinLeaf = 10,
  subset = NULL,
  weights = NULL,
  na.action = na.fail,
  Xscale = "Min-max",
  ...
)
}
\arguments{
\item{X}{An n by d numeric matrix (preferable) or data frame.}

\item{...}{Optional parameters to be passed to the low level function.}

\item{formula}{Object of class \code{formula} with a response describing the model to fit. If this is a data frame, it is taken as the model frame. (see \code{\link{model.frame}})}

\item{data}{Training data of class \code{data.frame} containing variables named in the formula. If \code{data} is missing it is obtained from the current environment by \code{formula}.}

\item{criteria}{The independence measurement criteria used for splitting the nodes. We provided seven criteria (default criteria='DCor'), for details see \code{\link{Split}}.}

\item{ntrees}{The number of trees in the forest (default 100).}

\item{storeOOB}{If TRUE then the samples omitted during the creation of a tree are stored as part of the tree (default TRUE).}

\item{replacement}{if TRUE then n samples are chosen, with replacement, from training data (default TRUE).}

\item{stratify}{If TRUE then class sample proportions are maintained during the random sampling. Ignored if replacement = FALSE (default TRUE).}

\item{ratOOB}{Ratio of 'out-of-bag' (default 1/3).}

\item{parallel}{Parallel computing or not (default TRUE).}

\item{numCores}{Number of cores to be used for parallel computing (default \code{Inf}).}

\item{MaxDepth}{The maximum depth of the tree (default \code{Inf}).}

\item{numNode}{Number of nodes that can be used by the tree (default \code{Inf}).}

\item{MinLeaf}{Minimum sample size for leaf nodes (Default 10).}

\item{subset}{An index vector indicating which rows should be used. (NOTE: If given, this argument must be named.)}

\item{weights}{Vector of non-negative observational weights; fractional weights are allowed (default NULL).}

\item{na.action}{A function to specify the action to be taken if NAs are found. (NOTE: If given, this argument must be named.)}

\item{Xscale}{Predictor standardization methods. " Min-max" (default), "Quantile", "No" denote Min-max transformation, Quantile transformation and No transformation respectively.}

\item{y}{A response vector of length n.}
}
\value{
An object of class SISforest Containing a list components:
\itemize{
\item{\code{call}: The original call to SISforest.}
\item{\code{terms}: An object of class \code{c("terms", "formula")} (see \code{\link{terms.object}}) summarizing the formula. Used by various methods, but typically not of direct relevance to users.}
\item{\code{data}: The list of data related parameters used to build the forest.}
\item{\code{tree}: The list of tree related parameters used to build the tree.}
\item{\code{forest}: The list of forest related parameters used to build the forest.}
\item{\code{VarImp}: Variable screening results for forest structure.
\itemize{
\item{\code{rank.SISforest}: Variable ranking results for sure independence screening forest.}
\item{\code{rank.SIStree}: Variable ranking results for each sure independence screening tree.}
\item{\code{SISforest}: The correlation coefficients of each variable with the response for sure independence screening forest.}
}}
}
}
\description{
The sure independence screening forest (SISforest) is an ensemble of multiple SIStrees generated by sample bagging. SISforest performs slightly better than SIStree, but consumes significantly more time.
}
\examples{
#simulation data with continuous response.
X = matrix(rnorm(100*1000), 100, 1000)
y = X[,1] + X[,2]^2 + 2*X[,4]*X[,5] + rnorm(100)
\dontrun{
sis = SISforest(X, y, parallel=FALSE)
# show the first 20 variables ranked by SISforest
sis$VarImp$rank.SISforest[1:20]
}

#real data with categorical response.
data(breast_cancer)

sis <- SISforest(as.factor(diagnosis) ~ ., breast_cancer[,-1],
       criteria='Gini', parallel=FALSE)
#show variables ranked by SISforest
sis$VarImp$rank.SISforest

}
\seealso{
\code{\link{SIStree}}
}
\author{
Yu Liu and Zhibo Cai
}
\keyword{forest}
