% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIStree.R
\name{SIStree}
\alias{SIStree}
\alias{SIStree.formula}
\alias{SIStree.default}
\title{Sure Independence Screening for Tree Structures}
\usage{
SIStree(X, ...)

\method{SIStree}{formula}(
  formula,
  data = NULL,
  criteria = c("DCor", "Pearson", "CvM", "MI", "HHG", "HSIC", "Gini")[1],
  MaxDepth = Inf,
  numNode = Inf,
  MinLeaf = 10,
  subset = NULL,
  weights = NULL,
  na.action = na.fail,
  Xscale = "Min-max",
  ...
)

\method{SIStree}{default}(
  X,
  y,
  criteria = c("DCor", "Pearson", "CvM", "MI", "HHG", "HSIC", "Gini")[1],
  MaxDepth = Inf,
  numNode = Inf,
  MinLeaf = 10,
  subset = NULL,
  weights = NULL,
  na.action = na.fail,
  Xscale = "Min-max",
  ...
)
}
\arguments{
\item{X}{An n by d numeric matrix (preferable) or data frame.}

\item{...}{Optional parameters to be passed to the low level function.}

\item{formula}{Object of class \code{formula} with a response describing the model to fit. If this is a data frame, it is taken as the model frame. (see \code{\link{model.frame}})}

\item{data}{Training data of class \code{data.frame} containing variables named in the formula. If \code{data} is missing it is obtained from the current environment by \code{formula}.}

\item{criteria}{The independence measurement criteria used for splitting the nodes. We provided seven criteria, and the default being distance correlation coefficient (criteria='DCor').}

\item{MaxDepth}{The maximum depth of the tree (default \code{Inf}).}

\item{numNode}{Number of nodes that can be used by the tree (default \code{Inf}).}

\item{MinLeaf}{Minimum sample size for leaf nodes (Default 10).}

\item{subset}{An index vector indicating which rows should be used. (NOTE: If given, this argument must be named.)}

\item{weights}{Vector of non-negative observational weights; fractional weights are allowed (default NULL).}

\item{na.action}{A function to specify the action to be taken if NAs are found. (NOTE: If given, this argument must be named.)}

\item{Xscale}{Predictor standardization methods. " Min-max" (default), "Quantile", "No" denote Min-max transformation, Quantile transformation and No transformation respectively.}

\item{y}{A response vector of length n.}
}
\value{
An object of class SIStree containing a list of components::
\itemize{
\item{\code{call}: The original call to SIStree.}
\item{\code{terms}: An object of class \code{c("terms", "formula")} (see \code{\link{terms.object}}) summarizing the formula. Used by various methods, but typically not of direct relevance to users.}
\item{\code{data}: The list of data related parameters used to build the tree.}
\item{\code{tree}: The list of tree related parameters used to build the tree.}
\item{\code{structure}: A set of tree structure data records.
\itemize{
\item{\code{nodeXIndx}: Training samples used for each node.}
\item{\code{nodeCutVar}: Splitting variables for each node.}
\item{\code{nodeCutValue}: Splitting Values for each node.}
\item{\code{childNode}: The number of child nodes after each splitting.}
\item{\code{nodeDepth}: The depth of the tree where each node is located.}
}}
\item{\code{VarImp}: Variable screening results for tree structure.
\itemize{
\item{\code{rank.SIS}: Variable ranking results for sure independence screening.}
\item{\code{rank.SIStree}: Variable ranking results for sure independence screening tree.}
\item{\code{rank.SIStree.layer}: Variable ranking results in each layer for sure independence screening tree.}
\item{\code{SIStree}: The correlation coefficients of each variable with the response for sure independence screening tree.}
\item{\code{SIStree.layer}: The correlation coefficients of each variable with the response in each layer for sure independence screening tree.}
\item{\code{splitVar}: The splitting variables for each splitting node.}
}}
}
}
\description{
Sure Independence Screening Tree (SIStree) by partitioning the sample into subsets consecutively, leading to a tree structure of sub-samples. While the splitting values are chosen by one of six independence measurement criteria.
}
\examples{
#simulation data with continuous response.
X = matrix(rnorm(100*1000), 100, 1000)
y = X[,1] + X[,2]^2 + 2*X[,4]*X[,5] + rnorm(100)

sis = SIStree(X, y)
# show the first 20 variables ranked by SIS
sis$VarImp$rank.SIS[1:20]
# show the first 20 variables ranked by SIStree
sis$VarImp$rank.SIStree[1:20]


#real data with categorical response.
data(breast_cancer)

sis <- SIStree(as.factor(diagnosis) ~ ., breast_cancer[,-1], criteria='Gini')
#show variables ranked by SIStree
sis$VarImp$rank.SIStree

}
\seealso{
\code{\link{SISforest}}
}
\author{
Yu Liu and Zhibo Cai
}
\keyword{tree}
